react-app: refactor VoteOptionIndicator out refs oursky/likedao#171
